Split RoundConfiguration into separate model file
