+++ frontend/src/app/models/round-configuration.model.ts
@@ -1,0 +1,12 @@
+export interface RoundConfiguration {
+  roundNumber: number;
+  themeId: string;
+  themeName: string;
+  questionTypes: {
+    type: string;
+    enabled: boolean;
+    questionCount: number;
+    maxAvailable: number;
+  }[];
+  totalQuestions: number;
+}
